Client: client side is probably finished (requires the server to test), all warnings were fixed

+++ STiVClient/Graphics/Shaders/PostProcessingManager.cpp
@@ -7,7 +7,7 @@ void PostProcessingManager::init()
     mShaderFromType.emplace(std::make_pair(MultitargetShaderType::Explosion, MultitargetShader("Content\\Shaders\\explosion.frag")));
 }
 
-void PostProcessingManager::addTargetToShaderProcessing(MultitargetShaderType shader, const sf::Vector2f& where, float time)
+void PostProcessingManager::addTargetToShaderProcessing(MultitargetShaderType shader, const sf::Vector2f& where, int time)
 {
     if (shader == MultitargetShaderType::None)
         return;
